requirements.txt
asyncpg==0.29.0
nats-py==2.6.0
redis==5.0.8
cryptography==42.0.8
